vim: Replace neosnippet with UltiSnips
The truth is neosnippet kind of sucked badly, hard to remember syntax,
awful need of hard tab characters, although kind of cool when it came to
integrate it with emmet (or anything tab-related). I'm willing to dig
deeper into vim in order to get UltiSnips to play nice with emmet and or
coc.nvim.

I feel like going the extra mile with UltiSnips is worth