zsh: Alias ssh to TERM=xterm-256color ssh
I use a different TERM database, which is usually not present in remote
systems, and that results in weird behavior, especially when opening
vim. Most systems I work on support 256color terminfo databases, so this
should be good. I'll dig further if I happen to get involved with a
system that doesn't support 256color. For now this is enough.