neovim: Fix markdown editing
Remove spelling for now because it seems to be causing a mess in my
windows any way I use it. Will eventually figure it out but for now
disable.

Also add frontmatter, which just makes sense because I either have it or
not and it's the only format that I seem to use at all for metadata