dwm: Add daemons for networkmanager and pulseaudio
First of all, dwmblocks executes every script in a timer, which is not
really ideal for some of the tasks I need to perform. Some of them I
don't care, really, like CPU usage, temperature and memory usage. Those
can be polled fine, but some of the elements of the bar *need* to be
immediate. These are network, caps lock and audio sink. I'll describe
those below.

Linux doesn't have the native capability of switching between pulseaudio
sinks for bluetooth, which means every time I need to hop into google
meet, Zoom, discord or whatever, no microphone is going to be detected
unless I manually switch from a2dp-sink to headset-head-unit. The
ach-pa-daemon signals to dwmblocks every time a change is detected using
the polybar script.

For networking, same applies, I can't be waiting for a polling script to
feed back the network status, so what I do is set a daemon to signal to
dwmblocks that something has changed over the network. It's actually a
pretty good solution. I'm very happy with it.