neovim, kitty: Map ctrl+i to F6 in kitty and C-i in vim
Might be a little confusing.

I recently read that it *might* be possible to distinguish between C-i
and <Tab> in vim. I also learned that they are completely different
keystrokes at least to kitty. However, I couldn't make that work inside
neovim.

I'm going back to the old wincent's trick, which is to make the
*terminal* think that C-i is something else, and then rebind that
something else (which is F6) to <C-i> inside neovim. That way, tab
mappings are separate from C-i mappings. Currently this is only relevant
in normal mode, because I'm not using C-i in insert or visual modes.
Now I could though.

Also, boy is this escape codes kitty has nonsense. I spent over an hour
and a half trying to figure out what was going on. It's documented
inside so I don't have to do this again. GOD.