zsh: Make config a little bit more universal
Up until this point I had been using macOS for around 2 years
exclusively, while that brought me a little joy in terms of
productivity, it didn't really when it comes to performance.

Anyways, the point is that my dotfiles are a mess right now, with
hardcoded paths that are macOS exclusive, a lot of macOS exclusive stuff
that I haven't even tested if it works in my Linux machine.

The point of this commit is to start moving towards a more universal
dotfiles repo. This includes:

- Properly using .zshenv
- Don't assume BSD/Linux for the ls command while still having colored
  output

In the future, I should be moving the "version manager" stuff out of
.zshrc because is a horrible mess and it's incredibly annoying to have
to keep that up to date.