polybar: Say no bt headphones if that's the case
I was previously just exiting, causing the `btsink` module to disappear
and never appear again. I'm pretty satisfied as it is now, although we
could get rid of the "No bt headphones" if there are no headphones, but
that would mean changing the daemon.

It's fine for now because I have reasonable space in my polybar