sxhkd: Use rofi instead of dmenu
dmenu is great but there's something about it that makes it lag when
selecting emoji. I don't wanna deal with that so I'm just switching to
rofi for now. This does not apply my workarounds for some executables
yet, but I will later.

Affected apps are:
- datagrip: Needs to set a Java env variable. I think I can get away
  with defining that variable outside rofi/dmenu
- discord: Needs --no-sandbox (actually I hate this) or audio crashes
- spotify: Needs to force scale factor to 1.5 on 4k monitors

Actually I'm just using Linux with a 4k monitor, so defaulting to 1.5x
viewport in spotify is fine for now. If I ever end up with some other
monitor (like an ultrawide or something), I'd make my new script decide,
but defaulting to 1.5 suffices for now