neovim: Replace telescope with fzf-lua
I actually didn't want to change this. Adding an external dependency
(skim in this case) is not something I'm particularly happy with.
However, telescope was underperforming for what I need to do:

- Some rg matches (that were shown in console) didn't show up in
  telescope live grep, which SUCKS
- Sometimes telescope left some artifacts in the screen. I could live
  with this, actually, but still an issue
- fzf (via skim) allows me to scroll, but telescope does not
- fzf opens my quickfix immediately. I'm sure I can do it in telescope
  but this works out of the box in fzf
- I had great trouble making telescope show hidden files but STILL
  respect gitignore. This is quite easy with rg + fzf

So, fzf via skim it is. It's not the prettiest UI (actually I think it's
a bit cluttered), but it sure as hell works, and works way better than
telescope.