x: Don't use GNOME keyring, load everything from x, launch dwm
I've moved to dwm for around 3 weeks to a month now and am pretty happy
with it. Since them I have fixed what is supposed to load when X is
started, I've set the default Gtk theme and I've stopped using the
default GNOME keyring. I'm not quite sure about the keyring yet though.
But my system is working fine for now