arch: Update package list
This is mostly to start using stow in the near future.

I have a pending wipe of my linux machine to fully set it up without
GNOME clutter. The general idea is to set up this repo so that it can be
automated, and part of it needs stow (at least for now)