installer: Try to guess install modules
I'm gonna start this scope about the installer.

- Ruby is an arbitrary choice, although not that much. It comes by
default in macOS and is easy to install in a Linux env. It also has
special syntax to parse output of system cmds.
- In this script we try to guess if we are in a Linux env or in a Mac
env, if we detect it, then we install the common module + the not so
common module (either arch or macos).
- In the future, I'd like to receive params to explicitly install some
modules instead of installing the whole thing at once