installer: Add us-altgr-intl keyboard layout in macos
That's not a default keyboard layout and is one that i'm very used to in
all my machines. I realize uploading a whole directory to github is very
far from ideal, but automating this is something that is crucial for me,
because I have had issues with Ukelele before. The bundle file is solid
right now (I believe I'm missing one key combination at this point).