zsh: Only export GPG tty variables in Linux
Haven't yet figured out how to do all this gpg thing in macOS. For the
record if I need to see this again:

> The purpose of those variables is to have the gpg agent manage my SSH
keys. If I do that, then I can use a PGP authentication key for ssh
auth, which would be great, because my key actually has an
authentication subkey.

But oh, well, haven't bothered to figure how to set up gpg and ssh in
macOS yet