kmonad: start config
I might be moving from interception-tools over to kmonad. General reason
is that interception-tools is way too low level for what I want and
programming timeouts and rollover is a fairly complicated topic. Kmonad
solves all my problems by giving me most of what I need from QMK at a
software level.

This configuration includes the 3 things I need the most on a keyboard:
- swap left alt and left super (not a big deal)
- tap caps for esc, hold for ctrl (this was being done by interception)
- spacefn while holding space, tap for normal space (not done before)

The last one is the big deal here and I presume it will be useful for
any OS from here on, which is GREAT