dwm: update bluetooth headphones script to POSIX
I've been waiting for this moment where I erase all legacy ruby code
from this repo and it's great.

This commit does break polybar, but if I'm ever in need of using polybar
again, this dwm-bluetooth-headphones script can be rearranged to be
useful in both context (dwmblocks + polybar). Until that time, this is
more than enough