hyprland: add conf
Might be moving out soon but still worth keeping around.

I really like the idea of this WM with all its graphic effects. It
doesn't feel sluggish at all, it's just a little buggy and I don't
really have the time to be debugging or contributing atm.

I'll probably check it out in a couple of months