kmonad: move left super and left alt to default positions
This is a hard hit in my muscle memory. I've been using the keyboard
this way for months now, but since coming back to Linux I'm having a
really hard time to see the benefits here.

This comment is going to be a little long but I feel like I have to
write it somewhere

So I've been using macOS since late 2018 after having been using linux
for many years. I had a little trouble at first adapting to cmd instead
of control but managed to get used to it. I then used this setup for a
couple of years until I bought a desktop linux PC because my mac was
underperforming.

I then went to great lengths in order to keep my workflow seamless
between my macs and my desktop. And I really needed it because I
switched machines a lot.

After that I moved on to 100% mac again for a couple months and now I'm
back to pretty much Linux.

I had a little trouble at first adapting to cmd instead of control but
managed to get used to it.
I then used this setup for a couple of years until I bought a desktop
linux PC because my mac was underperforming.

I then went to great lengths in order to keep my workflow seamless
between my macs and my desktop. And I really needed it because I
switched machines a lot.

After that I moved on to 100% mac again for a couple months and now I'm
back to pretty much Linux.

But the thing is that those lengths I traveled are no longer there. I
still use my mac from time to time without making too many muscle memory
mistakes and same goes for linux.

Over in the Linux side, though, I have taken a serious knockback on
keyboard efficiency because of this. You see, over on macos land, you
hit cmd and alt everywhere, because it's made for that. And I really
think it's more ergonomic than the PC setup with its overloaded control
key. However, the thing with this approach is that by swapping
alt/super, I'm making the alt key harder to reach, and that has some
effects, like not being able to use it as often in key bindings in the
terminal (alt-b, alt-f and so on and so forth). I do use super, for a
lot of things, but not as much as I should be using alt.

Or at least, that's what I believe, and I don't want to depend on these
kinds of bindings, really. Minimalism should be my goal here.

And this commit goes for that.