sway: set up screen locking
This config uses swayidle + swaylock in a classic fashion combining
`loginctl` to lock the session. Works pretty well with manual locking of
the screen (kind of like when you walk out of your workstation). It's no
perfect by any means and you only need to kill swaylock to gain access
back to the compositor.

As far as I know, there's not much to be done here in the current state
of affairs. There's a protocol for session locking:
https://wayland.app/protocols/ext-session-lock-v1  but it looks like it
doesn't have support in any compositor yet. For now I think it's fine
because I'm just using this PC at home.

However, in the future locking will be very important if I want to bring
a linux laptop to work. That would mean either waiting for this protocol
in sway or skipping sway altogether for just GNOME, until something
secure comes out.

I'm fine with this for the time being.