neovim: tweak cmp behavior
This deserves an explanation.

So I was using CR to accept a completion menu, regardless of the
existence of the selection itself. That meant that if nothing was
selected and I pressed CR, the completion would be accepted. It turns
out that's ok in 90% of cases despite being a little invasive. No
biggie, except that some sources are rather aggressive with completion,
like sumneko_lua, which try to populate with whatever they can, leading
to unwanted text being inserted.

I also commented about this issue in my latest tag:
https://github.com/anachronic/dotfiles/releases/tag/220924-working

So I set out to settle this behavior:

- CR will only accept if a completion is selected and will do whatever
it does if nothing is selected
- TAB will select next as usual, but if there's only one choice, it will
accept it and close the completion menu
- C-y will take the old CR role, meaning that it will eagerly accept
either the selected completion, or the first in the menu

There's a little downside to that, which is that Down will also accept
if there's only one choice. I don't use Down very often, so that
shouldn't be a problem. I guess I'm really lazy when it comes to
tweaking these things. W/e, if I need it I'll change it

This commit also replaces an `unpack` call with a `table.unpack`, which
is supposed to be how lua 5.4 expects me to use it