bin: add script to control volume
Ah, this is mainly needed because I believe some connection in my front
audio jack is messed up and it has to be connected in some way for the
OS to detect that there's something plugged in.

However, sound is actually fine, even if OS does not detect that there's
something in it. That makes me have to switch audio every time I
restart, which I don't mind for now.

The thing that was bugging me a lot is that i didn't have volume control
other than scrolling my mouse wheel over waybar, which is horrible. This
script will use the running sink instead of the default for volume