foot: rebind C-i to F6 globally
This is mostly not a problem because foot can differentiate between C-i
and Tab and so this has no downsides for terminal apps that use legacy
keycodes (or whatever it's called)

I didn't want to have to do this but I'm kinda forced to. I don't want
to go back to kitty and foot does not have tab support, which means that
I end up with a clusterfuck of a tree structure when trying to organize
all my terminals in one workspace

👆 I tried for around a week and it really doesn't work out for me. I'll
be sticking with tmux for the time being and I'll pray that we get
either a better multiplexer or foot supports tabs. I could go back to
kitty but I really rather not. Maybe I'll get into that in a blog post
or w/e