zsh: add function to change titlebar
Was considering using this for a sway-based workflow for terminals
instead of a tmux sessions-based. Sway's not really that good imo, the
tree structure, while understandable (with a cs degree anyway), is not
that great for a quick workflow, which is not a problem in tmux.

And so, I'm keeping this because foot doesn't change titles
automatically (fair enough, tbh) in case I ever need to