neovim: use ruff for python linting
I'd like for some way to select my linter that doesn't involve any menu.
Something like having a custom .flake8 or .ruff file.

Not working on any other python project that doesn't use ruff and I'm
confident that sensible projects will move to it so I'll keep it as
default for now.

If I ever need to work on a project that does not use ruff a possible
workaround is to use a precommit hook