neovim: add telescope ui select
This is intended for code action replacement.

Why? Because the plugin I currently use for that sometimes adds a narrow
window and wraps text around, making it harder for me to even decide
what actions are there.

This one, although it uses telescope, doesn't describe what actions do
and doesn't provide any diff, is very consistent on the underlying menu,
which is mostly what I need.

This commit does not set up the extension.