neovim: set up treesitter for python
I had been using the old syntax way because tree sitter for python was
hideous. It used to not indent anything well and folding was rather
bad. It seems to be a lot better now so I'll give it a try, especially
now that new syntax is being added, the current plugin I
use (python-syntax) does not have support for that