waybar: add battery module
I'm not exactly sure how this is gonna play with my desktop right now,
because it obviously does not have a battery and I would like it to not
show if I can avoid it.

If it turns out it can't be trusted, I'll have to do something about it,
like differentiating by hostnames, or something. Hopefully this works
out well without many changes